Decouple client and server
