+++ .gitignore
@@ -364,3 +364,4 @@ FodyWeavers.xsd
 
 *.o
 2Dminecraft
+*.exe
