+++ core/ChunkManager.cpp
