+++ core/ChunkManager.hpp
