+++ server/PerlinNoise.hpp
@@ -1,7 +1,7 @@
 #ifndef __PERLINNOISE_HPP__
 #define __PERLINNOISE_HPP__
 
-#include "common.hpp"
+#include "../core/common.hpp"
 
 float PerlinNoise(Vector3 Pos, int Octaves, float ConstFrequency);
 void SetSeed(int givenSeed);
